+++ R/options.R
@@ -247,6 +247,9 @@ test_options <- function(title, admin_password,
 #' - \code{left_margin} (set to \code{0L})
 #' - \code{right_margin} (set to \code{0L})
 #'
+#' Note that this will hide the admin panel.
+#' See \code{show_footer} for further information.
+#'
 #' @param content_background_colour
 #' (Character scalar)
 #' Background colour for the display element for the primary test content;
@@ -266,6 +269,11 @@ test_options <- function(title, admin_password,
 #' (Logical scalar)
 #' Whether the footer should be shown (typically contains a link to the
 #' admin panel).
+#' To show the admin panel manually, enter the Javascript console
+#' once the app is running
+#' (you'll need to be using a proper web browser, rather than RStudio)
+#' and enter the command \code{show_admin_panel()}.
+#' To hide it again, log out and then enter \code{hide_admin_panel()}.
 #'
 #' @param left_margin
 #' (Integerish scalar)
@@ -286,6 +294,8 @@ test_options <- function(title, admin_password,
 #' (Logical scalar)
 #' Whether to display a link for logging in to the admin panel
 #' (defaults to \code{TRUE}).
+#' Note that this admin panel will by default be invisible if \code{show_footer = FALSE},
+#' but this can be changed by calling the Javascript function \code{show_admin_panel()}.
 #'
 #' @return
 #' A list for passing to the \code{display} argument of \code{test_options}.